From backend-payment into backend

See merge request s12-fintech-finance-sub1/S12P21A602!25

+++ BE/rebirth/src/main/java/com/kkulmoo/rebirth/user/infrastrucutre/entity/UserEntity.java
@@ -29,7 +29,7 @@ public class UserEntity {
 	private Long userId;
 
 	@Column(name = "consumption_pattern_id")
-	private Byte consumptionPatternId;
+	private Short consumptionPatternId;
 
 	@Column(name = "user_name", nullable = false, length = 10)
 	private String userName;
